+++ example/android/app/src/main/java/com/bearblock/callx/MainActivity.kt
@@ -1,15 +1,15 @@
 package com.bearblock.callx
 
-import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
 import com.facebook.react.defaults.DefaultReactActivityDelegate
+import com.callx.CallxReactActivity
 
 /**
- * MainActivity - Callx will automatically detect and setup lock screen handling
+ * MainActivity - Extends CallxReactActivity for automatic lockscreen handling
  * No manual setup required!
  */
-class MainActivity : ReactActivity() {
+class MainActivity : CallxReactActivity() {
 
   /**
    * Returns the name of the main component registered from JavaScript. This is used to schedule